v0.7 beta

+++ src/main/java/me/marin/lockout/LockoutRunnable.java
@@ -5,7 +5,7 @@ import me.marin.lockout.server.LockoutServer;
 @FunctionalInterface
 public interface LockoutRunnable {
 
-    default void runTaskAfter(int ticks) {
+    default void runTaskAfter(long ticks) {
         LockoutServer.gameStartRunnables.put(this, ticks);
     }
     void run();
